:art: -Started norming raycasting file (it is not going to well thought)  -Added new file for better norming of raycasting for the textures (that implies changing the Makefile as well) :art: -Normed init_config which had a too long line
