+++ srcs/config/init_config.c
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/28 20:07:59 by ldutriez          #+#    #+#             */
-/*   Updated: 2022/06/15 18:29:04 by user42           ###   ########.fr       */
+/*   Updated: 2022/06/15 18:36:54 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ static void	init_texture(t_system *sys, char *line, t_int2 pos
 		pos.x++;
 	if (ft_is_valid_file_path(line + pos.x) == b_false)
 		return (add_error(parser, P_ERR_TEXTURE, pos.y, pos.x));
-	texture->img = mlx_xpm_file_to_image(sys->mlx.mlx, line + pos.x, &texture->size.x,
-			&texture->size.y);
+	texture->img = mlx_xpm_file_to_image(sys->mlx.mlx, line + pos.x,
+			&texture->size.x, &texture->size.y);
 	if (texture->img == NULL)
 		add_error(parser, P_ERR_TEXTURE, pos.y, pos.x);
 	texture->addr = mlx_get_data_addr(texture->img, &texture->bits_pxl,
